nav menu issue
Hey, Kevin!
It's Vukasin, from youtube. We talked about this issue with slide out menu, how it creates a bug, by expanding the menus width, and also, because of moving the container-all, it creates a horizontal scroll, and it messes up the layout. (Try it in mobile mode, and zoom out a little bit. You'll see that the layout is messed up. It pushes the .container-all almost off the screen)

This is my quick fix, on the spot. I'm sure it can be done much nicer.
 You don't have to pull it, but I just wanted to show you the difference. 
Notice how, in order to move close button, I had to give him translateX(678%). Very odd number.  (You have to do this, once you turn of the translateX for .container-all. For some reaseon, button stays at same place, so, I had to give him that coordinate )